Added cart api's

+++ e-commerce/src/main/java/com/project/e_commerce/dto/cart/CartItemResponse.java
@@ -12,5 +12,5 @@ public class CartItemResponse {
     private String productName;
     private BigDecimal price;
     private int quantity;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 }
